Made heavy revisions to add Linux support

+++ WinQuake/common.h
@@ -155,18 +155,18 @@ float MSG_ReadAngle (void);
 
 //============================================================================
 
-void Q_memset (void *dest, int fill, int count);
-void Q_memcpy (void *dest, const void *src, int count);
-int Q_memcmp (void *m1, void *m2, int count);
+void Q_memset (void *dest, int fill, size_t count);
+void Q_memcpy (void *dest, const void *src, size_t count);
+int Q_memcmp (void *m1, void *m2, size_t count);
 void Q_strcpy (char *dest, const char *src);
-void Q_strncpy (char *dest, const char *src, int count);
+void Q_strncpy (char *dest, const char *src, size_t count);
 int Q_strlen (const char *str);
 char *Q_strrchr (const char *s, char c);
 void Q_strcat (char *dest, const char *src);
 int Q_strcmp (const char *s1, const char *s2);
-int Q_strncmp (const char *s1, const char *s2, int count);
+int Q_strncmp (const char *s1, const char *s2, size_t count);
 int Q_strcasecmp (const char *s1, const char *s2);
-int Q_strncasecmp (const char *s1, const char *s2, int n);
+int Q_strncasecmp (const char *s1, const char *s2, size_t n);
 int	Q_atoi (const char *str);
 float Q_atof (const char *str);
 int q_vsnprintf(char* str, size_t size, const char* format, va_list args);
