+++ WinQuake/d_edge.cpp
@@ -197,7 +197,7 @@ void D_DrawSurfaces(void)
 			d_zistepv = s->d_zistepv;
 			d_ziorigin = s->d_ziorigin;
 
-			D_DrawSolidSurface(s, (int)s->data & 0xFF);
+            D_DrawSolidSurface(s, (size_t)s->data & 0xFF);
 			ResolveRenderer()->D_DrawZSpans(s->spans);
 		}
 	}
